fixed error

+++ PA4/src/parMatrix.cpp
@@ -450,7 +450,7 @@ int getIdRight( int myProcessor, int numProcessors )
     {
     int idToMyRight;
 
-    if( (myProcessor % sqrt(numProcessors)) < (int)(sqrt(numProcessors) - 1) )
+    if( (myProcessor % (int)sqrt(numProcessors)) < (int)(sqrt(numProcessors) - 1) )
         {
             idToMyRight = myProcessor -  ( sqrt(numProcessors) + 1 );
         }
